requirements.txt
accelerate
wandb
torchvision